refactor close frames
