+++ core/clientconn.go
@@ -5,44 +5,7 @@ import (
 	"time"
 
 	"github.com/OguzhanE/saltyrtc-server-go/pkg/base"
-	ws "github.com/gobwas/ws"
 )
-
-// CloseFrameNormalClosure //
-var CloseFrameNormalClosure = buildCloseFrame(base.CloseCodeNormalClosure, "")
-
-// CloseFrameGoingAway //
-var CloseFrameGoingAway = buildCloseFrame(base.CloseCodeGoingAway, "")
-
-// CloseFrameSubprotocolError //
-var CloseFrameSubprotocolError = buildCloseFrame(base.CloseCodeSubprotocolError, "")
-
-// CloseFramePathFullError //
-var CloseFramePathFullError = buildCloseFrame(base.CloseCodePathFullError, "")
-
-// CloseFrameProtocolError //
-var CloseFrameProtocolError = buildCloseFrame(base.CloseCodeProtocolError, "")
-
-// CloseFrameInternalError //
-var CloseFrameInternalError = buildCloseFrame(base.CloseCodeInternalError, "")
-
-// CloseFrameHandover //
-var CloseFrameHandover = buildCloseFrame(base.CloseCodeHandover, "")
-
-// CloseFrameDropByInitiator //
-var CloseFrameDropByInitiator = buildCloseFrame(base.CloseCodeDropByInitiator, "")
-
-// CloseFrameInitiatorCouldNotDecrypt //
-var CloseFrameInitiatorCouldNotDecrypt = buildCloseFrame(base.CloseCodeInitiatorCouldNotDecrypt, "")
-
-// CloseFrameNoSharedTasks //
-var CloseFrameNoSharedTasks = buildCloseFrame(base.CloseCodeNoSharedTasks, "")
-
-// CloseFrameInvalidKey //
-var CloseFrameInvalidKey = buildCloseFrame(base.CloseCodeInvalidKey, "")
-
-// CloseFrameTimeout //
-var CloseFrameTimeout = buildCloseFrame(base.CloseCodeTimeout, "")
 
 type ClientConn struct {
 	Conn      net.Conn
@@ -86,12 +49,4 @@ func (c ClientConn) Close(closeFrame []byte) error {
 		c.AliveStat = base.MarkAsDeath(c.AliveStat)
 	}
 	return err
-}
-
-func buildCloseFrame(code int, reason string) []byte {
-	return ws.MustCompileFrame(
-		ws.NewCloseFrame(ws.NewCloseFrameBody(
-			ws.StatusAbnormalClosure, reason,
-		)),
-	)
 }
