+++ core/conn.go
@@ -4,7 +4,46 @@ import (
 	"net"
 	"reflect"
 	"syscall"
+
+	"github.com/OguzhanE/saltyrtc-server-go/pkg/base"
+	ws "github.com/gobwas/ws"
 )
+
+// CloseFrameNormalClosure //
+var CloseFrameNormalClosure = buildCloseFrame(base.CloseCodeNormalClosure, "")
+
+// CloseFrameGoingAway //
+var CloseFrameGoingAway = buildCloseFrame(base.CloseCodeGoingAway, "")
+
+// CloseFrameSubprotocolError //
+var CloseFrameSubprotocolError = buildCloseFrame(base.CloseCodeSubprotocolError, "")
+
+// CloseFramePathFullError //
+var CloseFramePathFullError = buildCloseFrame(base.CloseCodePathFullError, "")
+
+// CloseFrameProtocolError //
+var CloseFrameProtocolError = buildCloseFrame(base.CloseCodeProtocolError, "")
+
+// CloseFrameInternalError //
+var CloseFrameInternalError = buildCloseFrame(base.CloseCodeInternalError, "")
+
+// CloseFrameHandover //
+var CloseFrameHandover = buildCloseFrame(base.CloseCodeHandover, "")
+
+// CloseFrameDropByInitiator //
+var CloseFrameDropByInitiator = buildCloseFrame(base.CloseCodeDropByInitiator, "")
+
+// CloseFrameInitiatorCouldNotDecrypt //
+var CloseFrameInitiatorCouldNotDecrypt = buildCloseFrame(base.CloseCodeInitiatorCouldNotDecrypt, "")
+
+// CloseFrameNoSharedTasks //
+var CloseFrameNoSharedTasks = buildCloseFrame(base.CloseCodeNoSharedTasks, "")
+
+// CloseFrameInvalidKey //
+var CloseFrameInvalidKey = buildCloseFrame(base.CloseCodeInvalidKey, "")
+
+// CloseFrameTimeout //
+var CloseFrameTimeout = buildCloseFrame(base.CloseCodeTimeout, "")
 
 // Conn ..
 type Conn struct {
@@ -31,4 +70,12 @@ func socketFD(conn net.Conn) int {
 	pfdVal := reflect.Indirect(fdVal).FieldByName("pfd")
 
 	return int(pfdVal.FieldByName("Sysfd").Int())
+}
+
+func buildCloseFrame(code int, reason string) []byte {
+	return ws.MustCompileFrame(
+		ws.NewCloseFrame(ws.NewCloseFrameBody(
+			ws.StatusAbnormalClosure, reason,
+		)),
+	)
 }
